# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PQ Verifiable Archive",
    "version": "2.0.0",
    "description": "Post-quantum document anchoring to Algorand mainnet with ML-DSA-65 (NIST FIPS-204) signatures. Pay $0.01 USDC per anchor via x402. Verify for free.",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "url": "https://github.com/m-reynaldo35/pq-verifiable-archive"
    }
  },
  "servers": [
    {
      "url": "https://pq-verifiable-archive-production.up.railway.app",
      "description": "Hosted API (pay-per-anchor)"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Service is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/anchor": {
      "post": {
        "summary": "Anchor a document hash to Algorand mainnet",
        "operationId": "anchorDocument",
        "description": "Requires an x402 payment of $0.01 USDC on Algorand mainnet. Send the request without a payment header first to receive the 402 payment-required instructions, then resend with the payment-signature header. Each payment header can be used once (reuse returns 409). Rate limited to 10 requests per minute per client.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "hash"
                ],
                "properties": {
                  "hash": {
                    "type": "string",
                    "pattern": "^[0-9a-f]{64}$",
                    "description": "SHA-256 hash of the document (64-char lowercase hex)",
                    "example": "d88c1a367ad6e6bc6c62909b557418a92f2b8c2e1e41e6f390756f4d973094cc"
                  },
                  "envelope_id": {
                    "type": "string",
                    "description": "Optional identifier for the document or signing envelope",
                    "pattern": "^[A-Za-z0-9._:-]{1,128}$"
                  },
                  "signers": {
                    "type": "array",
                    "description": "Optional signers (max 50). Recorded with signerSource 'requester-asserted': the bundle proves the requester claimed them, not that they signed.",
                    "items": {
                      "type": "object",
                      "required": [
                        "name",
                        "email",
                        "signedAt"
                      ],
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "signedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "maxItems": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document anchored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "algorandTxnId": {
                      "type": "string",
                      "description": "Algorand mainnet transaction ID"
                    },
                    "algorandRound": {
                      "type": "integer",
                      "description": "Confirmed block round"
                    },
                    "bundle": {
                      "type": "object",
                      "description": "Self-contained proof bundle signed with ML-DSA-65",
                      "properties": {
                        "protocol": {
                          "type": "string",
                          "example": "pqva/2"
                        },
                        "envelopeId": {
                          "type": "string"
                        },
                        "documentHash": {
                          "type": "string"
                        },
                        "batchId": {
                          "type": "string"
                        },
                        "merkleRoot": {
                          "type": "string",
                          "description": "SHA-256(0x00 || documentHash) for a single-document anchor"
                        },
                        "merkleProof": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "side": {
                                "type": "string",
                                "enum": [
                                  "left",
                                  "right"
                                ]
                              },
                              "hash": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "algorandTxnId": {
                          "type": "string"
                        },
                        "algorandRound": {
                          "type": "integer"
                        },
                        "blockTimestamp": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "stateProofRound": {
                          "type": "integer"
                        },
                        "issuerAddress": {
                          "type": "string",
                          "description": "Algorand address that sent the anchor txn"
                        },
                        "keyRegistrationTxnId": {
                          "type": "string"
                        },
                        "signers": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string"
                              },
                              "signedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        },
                        "signerSource": {
                          "type": "string",
                          "enum": [
                            "docusign-connect",
                            "requester-asserted"
                          ]
                        },
                        "algorithm": {
                          "type": "string",
                          "example": "ml-dsa-65"
                        },
                        "mldsaPublicKey": {
                          "type": "string"
                        },
                        "signature": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment required. Decode the `payment-required` response header (base64 JSON) for x402 payment instructions.",
            "headers": {
              "payment-required": {
                "description": "base64(JSON) x402 payment requirements",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (bad hash format or signer fields)"
          },
          "500": {
            "description": "Server error (Algorand anchor or signing failed)"
          },
          "409": {
            "description": "This payment header was already used"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "x402": []
          }
        ]
      }
    },
    "/api/verify": {
      "post": {
        "summary": "Verify a proof bundle (always free)",
        "operationId": "verifyBundle",
        "description": "Verifies a bundle against the issuer this server trusts: ML-DSA-65 signature and issuer key, document hash (if a PDF is supplied), Merkle inclusion, and the Algorand anchor txn (sender, round, note). Also reports, informationally, whether the indexer lists a covering state-proof txn (the Falcon-512 proof is not verified). The response includes documentChecked, signerSource and anchoredAt. No payment required.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "bundle"
                ],
                "properties": {
                  "bundle": {
                    "type": "string",
                    "format": "binary",
                    "description": "Proof bundle JSON file"
                  },
                  "pdf": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional original PDF to verify document hash"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "steps": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "passed": {
                            "type": "boolean"
                          },
                          "skipped": {
                            "type": "boolean"
                          },
                          "informational": {
                            "type": "boolean"
                          },
                          "detail": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "signers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "signedAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "payment-signature",
        "description": "x402 micropayment on Algorand. $0.01 USDC per anchor. Send without header first to receive 402 with payment instructions."
      }
    }
  }
}